Add new fields to CustomerDTO 4 Database purposes
I changed the behaviour for the DTO id, now it's no longer required.
This is because when saving to Firestore we don't want to add the
id to the document

I also added created, createdBy, updatedAt & email.

I also changed lastName to surname, birthDate to dateOfBirth
& type to be saved as int. This is because in our current Firestore database
we have this fields named like this. The Json Serializable package will
transform them as we need